V1.4 (Includes from V1.1 to V1.4) History ----------------------
SAPUME Patch  v.1.1

----------------------

 

- Differentiation in the lookupAttributeNames method between when the
query is done from the PolicyMap or from the Managed System.

There are 2 lists of different and configure  attributes for each case.
New key created in properties: sapume.general.managedSysAtts

 

- In suspend operation (disable user) besides updating the VALID_TO, the
user will be blocked.

And in the resume operation (user enable) in addition to updating the
VALID_TO, the user will be unlocked.

   

- Differentiation in the resetPassword method between when the
invocation to the operation is performed by a password change made by
the user

Or by a reset of the password made by an administrator. It is done by
querying the "Operation" parameter of the Request (SET_PASSWORD -
RESET_PASSWORD).

* When the resetPassword is invoked by a password change caused by the
user (setPassword) will always leave the operating account (no flag
involved)

* When the resetPassword is invoked by an administrative password change
(resetPassword) will be consulted, as was done, the flag in which the
value is mapped

Of the sapume.general.changePasswordAtNextLogon property of the
configuration file to see whether to leave the operative account or with
the pwd change flag at login

 

- Differentiation in the lookup method between roles and groups. As
there is no way to differentiate when you lookup from Group or Role
(since it does not arrive neither Clasification

       Nor want to use prefixes), it will look first as a Group and if
it is not found it will be searched as Role in SAPUME

----------------------

SAPUME Patch  v.1.2

----------------------

- It changes the logic for the differentiation in the lookup method
between roles and groups.

The extended attribute "GROUP_TYPE" of the ExtensibleObject of the
Request will be queried.

The accepted values ​​to set in the attribute of the ExtensibleObject
will be: GROUP and ROLE

In OpenIAM, this metadata must be inserted into the groovy script
associated with the SYSTEM pre-processor of the Lookup operation

If it can not be recovered, it will be searched in the 2: first as
ActivityGroup and if it is not found it will be searched as a Profile in
SAP

 

- Correction in the search method for the reconciliation of Groups and
Roles. The description of the Role / Group was not being stored
correctly

 

- Implementation for suspend () and resume () operations of the logic to
make the update of dates VALID_FROM and VALID_TO optional.

New key created in properties:
sapume.general.modifyDatesForEnableDisable. If it has True value, the
already implemented management will be performed to enable (VALID_TO =
9999-12-31)

And disable (VALID_TO = SYSDATE -1). In case of deactivating the flag
nothing will be done, so the account is only locked / unlocked as the
case may be.

 

- In the suspend () operation, a control has been added so that when a
user is created in SAP and it is trying to disable the same day, not
error. Only in that case,

Instead of assigning VALID_TO the value (CURRENT_DATE -1), it will be
assigned only the CURRENT_DATE.

 

----------------------

SAPUME Patch v.1.3

----------------------

 

- Definition of a Critical Section in the SAPUMEConnectorServiceImpl
class at the level of each method / operation, to limit possible
multithreading in the execution of the connector.

The critical section covers each method from beginning to end.

 

----------------------

SAPUME Patch v.1.4

----------------------

 

- - Bug fix in class SecuredString and that affected the establishment
of passwords with 'ñ' and other special characters (NO ANSI)